requirements.txt
streamlit
streamlit_navigation_bar
pandas
lxml
plotly-express
rapidfuzz